requirements.txt
decord==0.6.0
h5py==2.10.0
joblib==1.1.0
json5==0.9.6
matplotlib==3.5.1
opencv-python==4.2.0.34
Pillow==9.0.0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
tensorboard==2.8.0
tensorboardX==2.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.62.3